feat: add sound picker UI and AlarmSound enum model
- Implemented SoundPickerView for selecting and previewing default alarm sound
- Defined AlarmSound enum with id, displayName, and file metadata
- Added .caf files for low-buzz, high-buzz, and siren to Resources/Sounds
- Moved AlarmSound.swift from Resources to Models for proper separation of concerns
